+++ sdk/eventhub/azure-eventhub/CHANGELOG.md
@@ -4,8 +4,7 @@
 
 **New Features**
 
-- `azure.eventhub.exceptions.EventHubError` now inherits from `azure.core.exceptions.AzureError`.
-- `azure.eventhub.exceptions.OwnershipLostError` now inherits from `azure.eventhub.exceptions.EventHubError`.
+- `azure.eventhub.exceptions.EventHubError` and `azure.eventhub.exceptions.OwnershipLostError` now inherit from `azure.core.exceptions.AzureError`.
 
 **Bug fixes**